Add CHECKSUM file to the release page
This PR adds CHECKSUM file to the release page.
Currently, everything in dist/release/ on the build machine(travis) is
pushed to release page as in the `deploy` job config in travis configuration
So, this PR just loops around contents of dist/release and generates sha256
for each file in the folder and puts them in a file by name CHECKSUM in the
same path(dist/release/) expecting it to be uploaded to the release page.

fixes #736
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>